requirements.txt
fastapi
uvicorn
pydantic
aiofiles
psycopg2-binary